@@ -3,24 +3,24 @@
   "private": true,
   "name": "tscalc",
   "dependencies": {
-    "@swc/jest": "^0.1.2",
+    "@swc/jest": "^0.2.23",
     "@types/classnames": "^2.2.11",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^14.0.0",
-    "@types/react": "^17.0.2",
+    "@types/jest": "^29.2.3",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
     "@types/react-beautiful-dnd": "^13.0.0",
-    "@types/react-dom": "^17.0.1",
+    "@types/react-dom": "^18.0.9",
     "classnames": "^2.2.6",
     "fuse.js": "^6.4.6",
-    "gh-pages": "^3.1.0",
+    "gh-pages": "^4.0.0",
     "pako": "^2.0.0",
     "popper.js": "^1.16.1",
     "prettier": "^2.2.1",
-    "react": "^17.0.0",
+    "react": "^18.2.0",
     "react-beautiful-dnd": "^13.0.0",
-    "react-dom": "^17.0.0",
-    "react-scripts": "^4.0.3",
-    "recoil": "^0.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "recoil": "^0.7.6",
     "typescript": "^4.2.2"
   },
   "homepage": "https://terite.github.io/tscalc",
