@@ -1,5 +1,5 @@
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
 
 import { ErrorCatcher } from './components/ErrorCatcher';
@@ -9,12 +9,14 @@ import { ToastPortal } from './components/ToastPortal';
 
 import './index.css';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <ErrorCatcher>
     <RecoilRoot>
       <GameDataLoader child={(gameData) => <AppLoader gameData={gameData} />} />
       <ToastPortal />
     </RecoilRoot>
-  </ErrorCatcher>,
-  document.getElementById('root')
+  </ErrorCatcher>
 );
